Reorganizing files to make logical structure
